refactor: Profit 클래스 삭제
- 수익률 계산의 책임을 LottoResult가 가지도록 수정
