+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -52,4 +52,15 @@ class LottoResultTest {
 
         assertEquals(1, lottoResult.getLottoResult().get(Rank.FOURTH));
     }
+
+    @Test
+    @DisplayName("수익률 계산")
+    void calculate_profit_rate() {
+        LottoResult lottoResult = new LottoResult();
+        Money purchaseAmount = new Money(3000);
+
+        lottoResult.addMatchingCount(lottos, winningLotto);
+
+        assertEquals(lottoResult.calculateProfitRate(purchaseAmount), 676683.333, 0.01);
+    }
 }
